Android Gallery
This app displays thumbnail pictures in a grid layout.
Each thumbnail in the grid view supports two kinds of functionality.
First, a simple click on a thumbnail brings up a new activity that
shows the entire picture of the selected car on the entire device
display. The user can return to the grid view by selecting the “back”
soft button on the phone.
A long click on any thumbnail will bring up a “context menu” showing
the following two options for the car under consideration: (1) View the
entire picture (similar to a simple click); (2) Show the official web
page of the selected thumbnail in a new activity